# Customize the PNG encoder