--- a/lib/srv/desktop/rdp/rdpclient/client_common.go
+++ b/lib/srv/desktop/rdp/rdpclient/client_common.go
@@ -19,6 +19,7 @@ package rdpclient
 
 import (
 	"context"
+	"image/png"
 
 	"github.com/gravitational/teleport/lib/srv/desktop/tdp"
 	"github.com/gravitational/trace"
@@ -35,13 +36,12 @@ type Config struct {
 	// AuthorizeFn is called to authorize a user connecting to a Windows desktop.
 	AuthorizeFn func(login string) error
 
-	// TODO(zmb3): replace these callbacks with a tdp.Conn
+	// Conn handles TDP messages between Windows Desktop Service
+	// and a Teleport Proxy.
+	Conn *tdp.Conn
 
-	// InputMessage is called to receive a message from the client for the RDP
-	// server. This function should block until there is a message.
-	InputMessage func() (tdp.Message, error)
-	// OutputMessage is called to send a message from RDP server to the client.
-	OutputMessage func(tdp.Message) error
+	// Encoder is an optional override for PNG encoding.
+	Encoder *png.Encoder
 
 	// Log is the logger for status messages.
 	Log logrus.FieldLogger
@@ -58,14 +58,14 @@ func (c *Config) checkAndSetDefaults() error {
 	if c.GenerateUserCert == nil {
 		return trace.BadParameter("missing GenerateUserCert in rdpclient.Config")
 	}
-	if c.InputMessage == nil {
-		return trace.BadParameter("missing InputMessage in rdpclient.Config")
-	}
-	if c.OutputMessage == nil {
-		return trace.BadParameter("missing OutputMessage in rdpclient.Config")
+	if c.Conn == nil {
+		return trace.BadParameter("missing Conn in rdpclient.Config")
 	}
 	if c.AuthorizeFn == nil {
 		return trace.BadParameter("missing AuthorizeFn in rdpclient.Config")
+	}
+	if c.Encoder == nil {
+		c.Encoder = tdp.PNGEncoder()
 	}
 	if c.Log == nil {
 		c.Log = logrus.New()