… lib/srv/desktop/tdp/testdata/png_frames.json