Instead of using the default, allow for providing a custom `png.Encoder`. Set `BestSpeed` for compression and use a pool in order to reuse buffers.

Also add some sample PNG frames and a benchmark for encoding them. This results in ~ 70% reduction in encoding time and > 99% reduction in memory allocations.

    name         old time/op    new time/op
    EncodePNG-8     107µs ± 0%      30µs ± 0%

    name         old alloc/op   new alloc/op
    EncodePNG-8     850kB ± 0%       1kB ± 0%

    name         old allocs/op  new allocs/op
    EncodePNG-8      42.0 ± 0%      13.0 ± 0%

Updates #8742